+++ crates/api/src/unit.rs
@@ -14,6 +14,8 @@ mod partial_ord;
 mod reducible;
 mod term_reducing;
 mod to_reduced;
+#[cfg(feature = "v2")]
+mod v2;
 
 #[allow(clippy::module_name_repetitions)]
 mod ucum_unit;
@@ -48,7 +50,7 @@ pub const UNITY: Unit = Unit {
 )]
 #[derive(Clone, Debug)]
 pub struct Unit {
-    terms: Cow<'static, [Term]>,
+    pub(crate) terms: Cow<'static, [Term]>,
 }
 
 /// A `Unit` is the piece of data that represents a *valid* UCUM unit or