+++ crates/api/CHANGELOG.md
@@ -18,11 +18,15 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - PLCC-287: `impl num_traits::NumCast for Measurement`.
 - PLCC-287: `impl num_traits::Pow<i32> for Measurement`, `Unit`, `Term`.
 - PLCC-287: `impl std::ops::Neg for Measurement`.
-- NAUM-5: Added `crate::Term::as_cow_str()`.
 - NAUM-4: Derive `PartialOrd` for `Composition`
 - NAUM-4: `impl From<Dimension> for Composition`
 - NAUM-4: Derive `Hash` for `Property`
 - NAUM-4: `impl PartialOrd for Term`
+- NAUM-5: Added `crate::Term::as_cow_str()`.
+- NAUM-6
+  - Add new `"v2"` feature.
+  - Add `v2::convert::ToFraction` trait.
+  - Implemented `v2::convert::ToFraction` trait for `Unit`, `Measurement`.
 - Added `Unit::into_terms()` for cases where you only need the `Term`s of the `Unit`.
 - Added `unit` constant: `UNITY`
 - Added `term` constants: `UNITY`, `UNITY_ARRAY`, and `UNITY_ARRAY_REF`.
@@ -51,6 +55,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   - `ucum_symbol`
 - (Internal) Moved `crate::parser` to `crate::unit::parser`.
 - NAUM-5: (Internal) `Display` implementation for `Unit` now uses `Term::as_cow_str()` logic.
+- NAUM-6: Updated `AsFraction::as_fraction()` implementation to iterate through `Term`s once instead
+  of twice for building the numerator and denominator.
 
 ### Deprecated